requirements.txt
langchain
langchain-core
langchain-groq
langchain-community
python-dotenv
streamlit